% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f24_modified_1.0.r
\name{f24}
\alias{f24}
\title{Old version f24 function using fft}
\usage{
f24(x, offset = 0, nf24 = 2)
}
\arguments{
\item{x}{the data of time series}

\item{offset}{offset of time points}

\item{nf24}{nb of replicates}
}
\description{
This function tests if the data is 24h rhythmic
}
\author{
Felix Naef
}
