% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f24_modified_1.0.r
\name{f24_R2_ls}
\alias{f24_R2_ls}
\title{latest version f24 function using lm}
\usage{
f24_R2_ls(x, t = 2 * (0:(length(x) - 1)), period = 24, offset = 0)
}
\arguments{
\item{x}{the data of time series}

\item{t}{the corresponding time points}

\item{period}{the period to test}

\item{offset}{the offset of time points (not used)}
}
\value{
A vector of nb of time points, mean of data, amplitude and relative amplitude of fitted result,  phase and pvalue
}
\description{
This function tests if the data is 24h rhythmic
}
\details{
the function can deal with missing time points
if there are less 4 time points, no amplitude, relative amplitude, phase and pvalue were calculated
}
\seealso{
\code{lm}\
}
\author{
Jingkui Wang
}
