% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f24_modified_1.0.r
\name{qvals}
\alias{qvals}
\title{function to compute q value}
\usage{
qvals(pv)
}
\arguments{
\item{pv}{a vector of p values}
}
\value{
A vector of q values
}
\description{
This function implement the Benjamin-Hochberg (BH) method for q value calculation
}
\details{
the function can deal with missing time points
}
\author{
Laura Symul
}
